% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imimp.R
\name{imimp}
\alias{imimp}
\title{Refines image imputation with multiple imputation and conditional density estimation}
\usage{
imimp(
  yhat,
  calibYhat,
  calibY,
  cdeType = c("simple", "local", "bySample"),
  sampleID = NULL,
  n.mi = 10,
  x.mar = 50,
  y.mar = 1000
)
}
\arguments{
\item{yhat}{Predicted value for testing data.}

\item{calibYhat}{Predicted value for calibration data. For method `bySample`, it should be a vector of each training slide predicted value predicted by the other training slides.}

\item{calibY}{True value for calibration data.}

\item{cdeType}{Different types of conditional density estimation with calibration data, conditioning on the predicted value of predicted value of testing data.}

\item{sampleID}{(optional) Sample ID of each cell, only need to be provided for method `bySample`. It should be a list containing two vectors: the first for `calibY`, the second for `Yhat`.}

\item{n.mi}{(optional) Number of multiple imputations. Default to be 10.}

\item{x.mar}{(optional) Number of bins for conditional density estimation. Default to be 50.}

\item{y.mar}{(optional) umber of bins for conditional density estimation. Default to be 1000.}
}
\value{
\item{resmat}{The matrix of multiple imputations. The number of rows is equal to the number of cells, and the number of columns is the number of multiple imputation.}
}
\description{
Refines image imputation with multiple imputation and conditional density estimation
}
\examples{
# create mock data for different types of conditional density estimation
## method `simple` and `local`
yhat_sl <- rnorm(1000, sd=20)
calibY_sl <- runif(50, min=-5, max=5)
calibYhat_sl <- runif(50, min=-5, max=5)
test.imp.fun <- imimp(yhat_sl, calibYhat_sl, calibY_sl, cdeType = "simple", x.mar = 50, y.mar=100)
test.imp.fun <- imimp(yhat_sl, calibYhat_sl, calibY_sl, cdeType = "local", x.mar = 50, y.mar=100)
## method `bySample`
yhat_b <- rnorm(600)
calibY_b <- runif(700, min=-5, max=5)
calibYhat_b <- runif(700, min=-5, max=5)
sample_b <- list(rep(state.name[7:16], each=70), rep(state.name[1:6], each=100))
test.imp.fun <- imimp(yhat_b, calibYhat_b, calibY_b, cdeType = "bySample", 
                     x.mar = 50, y.mar=100, sampleID=sample_b)

}
