% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDE.R
\name{CDE}
\alias{CDE}
\title{Function for conditional sampling}
\usage{
CDE(Ytrue, Yhat, Ytest, xmar = 50, ymar = 100)
}
\arguments{
\item{Ytrue}{True value of calibration data}

\item{Yhat}{Predicted value of calibration data}

\item{Ytest}{Predicted value of testing data}

\item{xmar}{Number of bins for conditional density estimation}

\item{ymar}{Number of bins for conditional density estimation}
}
\description{
Function for conditional sampling
}
